Checkpoint: failed approach to overrides for parameterized types
Turns out that the type of field is not sufficient to determine how it
gets overridden. Rather, we need to use something like the full
context of the field, including it's containing type. I'll work that
next, but am recording the deadend here, since most of the changes
here will be preserved.