Add support for overrides in parameterized types
With a few examples.

I'm having trouble deciding what the "right" definitions and APIs are
for overrides in this more general case with parameters. The code in
this commit compiles and runs, but it's not clear it's a the right
solution. In particular, there is no way to pass overrides into
generators for types determined by type parameters. Adding support for
wildcards will motivate a solution to that problem, but I doubt it
will be obvious it's the "right" solution.